refactor: relocate dedupe configuration
