+++ config/config.go
@@ -2,23 +2,15 @@
 package config
 
 import (
-	"crypto/tls"
 	"errors"
 	"fmt"
 	"strings"
-	"time"
 
 	"github.com/confluentinc/confluent-kafka-go/v2/kafka"
-	"github.com/mitchellh/mapstructure"
-	"github.com/redis/go-redis/extra/redisotel/v9"
-	"github.com/redis/go-redis/v9"
 	"github.com/spf13/pflag"
 	"github.com/spf13/viper"
 	"github.com/thmeitz/ksqldb-go/net"
 
-	"github.com/openmeterio/openmeter/internal/dedupe/memorydedupe"
-	"github.com/openmeterio/openmeter/internal/dedupe/redisdedupe"
-	"github.com/openmeterio/openmeter/internal/ingest"
 	"github.com/openmeterio/openmeter/pkg/models"
 )
 
@@ -92,11 +84,11 @@ func (c Configuration) Validate() error {
 	}
 
 	if err := c.Dedupe.Validate(); err != nil {
-		return err
+		return fmt.Errorf("dedupe: %w", err)
 	}
 
 	if err := c.Telemetry.Validate(); err != nil {
-		return err
+		return fmt.Errorf("telemetry: %w", err)
 	}
 
 	for _, m := range c.Meters {
@@ -302,209 +294,6 @@ func (c KafkaSinkClickhouseConfiguration) Validate() error {
 	return nil
 }
 
-// Requires [mapstructurex.MapDecoderHookFunc] to be high up in the decode hook chain.
-type DedupeConfiguration struct {
-	Enabled bool
-
-	DedupeDriverConfiguration
-}
-
-func (c DedupeConfiguration) NewDeduplicator() (ingest.Deduplicator, error) {
-	if !c.Enabled {
-		return nil, errors.New("dedupe: disabled")
-	}
-
-	if c.DedupeDriverConfiguration == nil {
-		return nil, errors.New("dedupe: missing driver configuration")
-	}
-
-	return c.DedupeDriverConfiguration.NewDeduplicator()
-}
-
-func (c DedupeConfiguration) Validate() error {
-	if !c.Enabled {
-		return nil
-	}
-
-	if c.DedupeDriverConfiguration == nil {
-		return errors.New("dedupe: missing driver configuration")
-	}
-
-	if err := c.DedupeDriverConfiguration.Validate(); err != nil {
-		return fmt.Errorf("dedupe: %w", err)
-	}
-
-	return nil
-}
-
-type rawDedupeConfiguration struct {
-	Enabled bool
-	Driver  string
-	Config  map[string]any
-}
-
-func (c *DedupeConfiguration) DecodeMap(v map[string]any) error {
-	var rawConfig rawDedupeConfiguration
-
-	err := mapstructure.Decode(v, &rawConfig)
-	if err != nil {
-		return err
-	}
-
-	c.Enabled = rawConfig.Enabled
-
-	// Deduplication is disabled and not configured, so skip further decoding
-	if !c.Enabled && rawConfig.Driver == "" {
-		return nil
-	}
-
-	switch rawConfig.Driver {
-	case "memory":
-		var driverConfig DedupeDriverMemoryConfiguration
-
-		err := mapstructure.Decode(rawConfig.Config, &driverConfig)
-		if err != nil {
-			return fmt.Errorf("dedupe: decoding memory driver config: %w", err)
-		}
-
-		c.DedupeDriverConfiguration = driverConfig
-
-	case "redis":
-		var driverConfig DedupeDriverRedisConfiguration
-
-		decoder, err := mapstructure.NewDecoder(&mapstructure.DecoderConfig{
-			Metadata:         nil,
-			Result:           &driverConfig,
-			WeaklyTypedInput: true,
-			DecodeHook: mapstructure.ComposeDecodeHookFunc(
-				mapstructure.StringToTimeDurationHookFunc(),
-			),
-		})
-		if err != nil {
-			return fmt.Errorf("dedupe: creating decoder: %w", err)
-		}
-
-		err = decoder.Decode(rawConfig.Config)
-		if err != nil {
-			return fmt.Errorf("dedupe: decoding redis driver config: %w", err)
-		}
-
-		c.DedupeDriverConfiguration = driverConfig
-
-	case "":
-		return errors.New("dedupe: missing driver")
-
-	default:
-		return fmt.Errorf("dedupe: unknown driver: %s", rawConfig.Driver)
-	}
-
-	return nil
-}
-
-type DedupeDriverConfiguration interface {
-	NewDeduplicator() (ingest.Deduplicator, error)
-	Validate() error
-}
-
-// Dedupe memory driver configuration
-type DedupeDriverMemoryConfiguration struct {
-	Enabled bool
-	Size    int
-}
-
-func (c DedupeDriverMemoryConfiguration) NewDeduplicator() (ingest.Deduplicator, error) {
-	return memorydedupe.NewDeduplicator(c.Size)
-}
-
-func (c DedupeDriverMemoryConfiguration) Validate() error {
-	if c.Size == 0 {
-		return errors.New("memory: size is required")
-	}
-
-	return nil
-}
-
-// Dedupe redis driver configuration
-type DedupeDriverRedisConfiguration struct {
-	Address    string
-	Database   int
-	Username   string
-	Password   string
-	Expiration time.Duration
-	Sentinel   struct {
-		Enabled    bool
-		MasterName string
-	}
-	TLS struct {
-		Enabled            bool
-		InsecureSkipVerify bool
-	}
-}
-
-func (c DedupeDriverRedisConfiguration) NewDeduplicator() (ingest.Deduplicator, error) {
-	var tlsConfig *tls.Config
-
-	if c.TLS.Enabled {
-		tlsConfig = &tls.Config{
-			InsecureSkipVerify: c.TLS.InsecureSkipVerify,
-		}
-	}
-
-	var redisClient *redis.Client
-
-	if c.Sentinel.Enabled {
-		redisClient = redis.NewFailoverClient(&redis.FailoverOptions{
-			MasterName:    c.Sentinel.MasterName,
-			SentinelAddrs: []string{c.Address},
-			DB:            c.Database,
-			Username:      c.Username,
-			Password:      c.Password,
-			TLSConfig:     tlsConfig,
-		})
-	} else {
-		redisClient = redis.NewClient(&redis.Options{
-			Addr:      c.Address,
-			DB:        c.Database,
-			Username:  c.Username,
-			Password:  c.Password,
-			TLSConfig: tlsConfig,
-		})
-	}
-
-	// Enable tracing
-	// TODO: use configured tracer provider
-	if err := redisotel.InstrumentTracing(redisClient); err != nil {
-		return nil, err
-	}
-
-	// Enable metrics
-	// TODO: use configured tracer provider
-	if err := redisotel.InstrumentMetrics(redisClient); err != nil {
-		return nil, err
-	}
-
-	// TODO: close redis client when shutting down
-	// TODO: register health check for redis
-	return redisdedupe.Deduplicator{
-		Redis:      redisClient,
-		Expiration: c.Expiration,
-	}, nil
-}
-
-func (c DedupeDriverRedisConfiguration) Validate() error {
-	if c.Address == "" {
-		return errors.New("redis: address is required")
-	}
-
-	if c.Sentinel.Enabled {
-		if c.Sentinel.MasterName == "" {
-			return errors.New("redis: sentinel: master name is required")
-		}
-	}
-
-	return nil
-}
-
 // Configure configures some defaults in the Viper instance.
 func Configure(v *viper.Viper, flags *pflag.FlagSet) {
 	// Viper settings
@@ -568,20 +357,5 @@ func Configure(v *viper.Viper, flags *pflag.FlagSet) {
 	v.SetDefault("sink.kafkaConnect.clickhouse.username", "default")
 	v.SetDefault("sink.kafkaConnect.clickhouse.password", "")
 
-	v.SetDefault("dedupe.enabled", false)
-	v.SetDefault("dedupe.driver", "memory")
-
-	// Dedupe Memory configuration
-	v.SetDefault("dedupe.config.size", 128)
-
-	// Dedupe Redis configuration
-	v.SetDefault("dedupe.config.address", "127.0.0.1:6379")
-	v.SetDefault("dedupe.config.database", 0)
-	v.SetDefault("dedupe.config.username", "")
-	v.SetDefault("dedupe.config.password", "")
-	v.SetDefault("dedupe.config.expiration", "24h")
-	v.SetDefault("dedupe.config.sentinel.enabled", false)
-	v.SetDefault("dedupe.config.sentinel.masterName", "")
-	v.SetDefault("dedupe.config.tls.enabled", false)
-	v.SetDefault("dedupe.config.tls.insecureSkipVerify", false)
+	configureDedupe(v)
 }
